split test module
